data_science/blood cell analysis/requirements.txt
# Datenverarbeitung & EDA
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
Pillow>=10.0

# Machine Learning
scikit-learn>=1.3

# Jupyter
jupyter>=1.0
ipykernel>=6.0
ipywidgets>=8.0

# SQL
# sqlite3 ist in Python eingebaut – kein extra Package nötig

# Optional: Deep Learning (nur wenn CNN geplant)
# tensorflow>=2.13
# keras>=2.13

# Optional: Kaggle API (zum Herunterladen des Datasets)
# kaggle>=1.5
